Removing the static classes from the main Async class and making them their own for grpc
Removing the static classes from the main Async class and making them their own for grpc so we can use them in the proto file
