+++ src/project/networkapi/AsyncJobResponse.java
@@ -1,0 +1,7 @@
+package project.networkapi;
+
+public interface AsyncJobResponse {
+    RequestStatus getStatus();
+    String getJobId();
+    String getMessage();
+}
