+++ src/project/networkapi/AsyncUserComputeAPI.java
@@ -559,132 +559,6 @@ public class AsyncUserComputeAPI implements UserComputeAPI {
 		}
 	}
 
-	// New response classes needed for async operations
-	public interface AsyncJobResponse {
-		RequestStatus getStatus();
-		String getJobId();
-		String getMessage();
-	}
-
-	public interface JobResultResponse {
-		RequestStatus getStatus();
-		String getResultData();
-		String getMessage();
-	}
-
-	public interface JobListResponse {
-		RequestStatus getStatus();
-		List<JobSummary> getJobs();
-		String getMessage();
-	}
-
-	public static class BasicAsyncJobResponse implements AsyncJobResponse {
-		private final RequestStatus status;
-		private final String jobId;
-		private final String message;
-
-		public BasicAsyncJobResponse(RequestStatus status, String jobId, String message) {
-			this.status = status;
-			this.jobId = jobId;
-			this.message = message;
-		}
-
-		@Override 
-		public RequestStatus getStatus() {
-			return status; 
-		}
-		@Override 
-		public String getJobId() { 
-			return jobId; 
-		}
-		@Override 
-		public String getMessage() { 
-			return message;
-		}
-	}
-
-	public static class BasicJobResultResponse implements JobResultResponse {
-		private final RequestStatus status;
-		private final String resultData;
-		private final String message;
-
-		public BasicJobResultResponse(RequestStatus status, String resultData, String message) {
-			this.status = status;
-			this.resultData = resultData;
-			this.message = message;
-		}
-
-		@Override public RequestStatus getStatus() { 
-			return status; 
-		}
-		@Override 
-		public String getResultData() { 
-			return resultData; 
-		}
-		@Override 
-		public String getMessage() { 
-			return message; 
-		}
-	}
-
-	public static class BasicJobListResponse implements JobListResponse {
-		private final RequestStatus status;
-		private final List<JobSummary> jobs;
-		private final String message;
-
-		public BasicJobListResponse(RequestStatus status, List<JobSummary> jobs, String message) {
-			this.status = status;
-			this.jobs = jobs;
-			this.message = message;
-		}
-
-		@Override 
-		public RequestStatus getStatus() { 
-			return status; 
-		}
-		@Override
-		public List<JobSummary> getJobs() { 
-			return jobs; 
-		}
-		@Override 
-		public String getMessage() { 
-			return message; 
-		}
-	}
-
-	public static class JobSummary {
-		private final String jobId;
-		private final CompletionStatus status;
-		private final int progress;
-		private final String outputDestination;
-		private final String message;
-
-		public JobSummary(String jobId, CompletionStatus status, int progress, 
-				String outputDestination, String message) {
-			this.jobId = jobId;
-			this.status = status;
-			this.progress = progress;
-			this.outputDestination = outputDestination;
-			this.message = message;
-		}
-
-		public String getJobId() { 
-			return jobId; 
-		}
-		public CompletionStatus getStatus() { 
-			return status; 
-		}
-		public int getProgress() { 
-			return progress; 
-		}
-		public String getOutputDestination() { 
-			return outputDestination; 
-		}
-		public String getMessage() { 
-			return message; 
-		}
-	}
-
 	// Getters for current configuration
 	public String getCurrentInputSource() { 
 		return currentInputSource;
