+++ src/project/networkapi/JobListResponse.java
@@ -1,0 +1,9 @@
+package project.networkapi;
+
+import java.util.List;
+
+public interface JobListResponse {
+	RequestStatus getStatus();
+	List<JobSummary> getJobs();
+	String getMessage();
+}
