+++ src/project/networkapi/JobResultResponse.java
@@ -1,0 +1,7 @@
+package project.networkapi;
+
+public interface JobResultResponse {
+		RequestStatus getStatus();
+		String getResultData();
+		String getMessage();
+	}
